refactored appointment_editor.dart
