--- .vscode/launch.json
@@ -1,29 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "final_project",
-            "cwd": "final_project",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "debug"
-        },
-        {
-            "name": "final_project (profile mode)",
-            "cwd": "final_project",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "debug"
-        },
-        {
-            "name": "final_project (release mode)",
-            "cwd": "final_project",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "debug"
-        }
-    ]
-}
