+++ final_project/lib/pages/user_calendar_page.dart
@@ -8,7 +8,7 @@ import 'package:syncfusion_flutter_calendar/calendar.dart';
 import 'package:final_project/objects/event.dart';
 import 'package:final_project/objects/group.dart';
 import 'package:final_project/objects/lake_nixon_event.dart';
-import 'package:final_project/appointment_editor/appointment_editor.dart';
+import 'package:final_project/pages/appointment_editor.dart';
 
 List<LakeNixonEvent> appointments = <LakeNixonEvent>[];
