+++ final_project/lib/widgets/appt_editor_select_rule_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:syncfusion_flutter_calendar/calendar.dart';
 
-import 'package:final_project/appointment_editor/custom_rule.dart';
+import 'package:final_project/widgets/appt_editor_custom_rule.dart';
 
 class SelectRuleDialog extends StatefulWidget {
   const SelectRuleDialog(
